fix: 게임 구매 시 돈이 추가되는 오류 수정

+++ src/backend/payment/src/main/java/com/steam/payment/entity/User.java
@@ -61,7 +61,7 @@ public class User {
         this.money += value;
     }
 
-    public void substractMoney(Integer value) {
+    public void substractMoney(Double value) {
         this.money -= value;
     }
 }
